Added a get info function
Gets a single line of information from the user.
Opens a float, focuses it, enters insert mode, and registers several
commands and keybinds to the new floating buffer.
Behavior is that when enter is pressed, information is passed to
callafter.
When escape is pressed, float is closed, insert mode is exited, and
function exits.
When text is changed in the buffer in insert mode the entered text is
printed.
This is not the final behavior.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>