Improved user commands
Currently EditTemplate and FromTemplate are completed as intended.
NewTemplate needs to be modified as it currently operates the same as
EditTemplate but should act as follows.
Opens an entry buffer, auto complete on all directories in the templates
folder as the user enters a path, on enter new template is created at
path specified and is opened in new buffer.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>