Added edit_template
Takes a template path and opens it for editing.
Closes the path upon saving buffer or file and returns to the previous
buffer. Returns to the intro screen if no buffer exists to return to.
Intro screen can be changed to whatever functionaliy is desired by
modifying the `intro` config variable.
Default value is vim.intro or :Alpha depending on if alpha.nvim is
installed. Will all options for startify and snacks.nvim at some point.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>