Added new_template module
Allows creating a new template in the configured templates directory.
Prompts the user for the file name. Float is not finished but function
works as intended. Will need to test it for crashes later but for now
under normal cases does not crash.
Creates a new template with the possibility to clobber old templates.
Will modify new_template to act like edit_template if the file being
'created' already exists.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>