Cleaned up new_template and added ability to change callback
Cleaned up code for new_template function. Removed and reformatted old
code to make it cleaner. Added correct title banner.
Added ability to change callback so that an arbitrary function can be
called instead of the default function. If opts.callback is not set the
default is run.
NOTE: There is a chance that this impl could cause a preset
opts.callback which was not set for specific use here to be called.
I don't see this happening but it might.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>