Added from_template
Prompts the user to select a template the same way edit_template does.
Then gets a path and file name from the user and calls
util.file_from_template with the choosen template and file path as args.
Currently does not get the correct root dir for selecting the file path.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>