Updated get_name with new usecase
Previous use case for get_name left several unneeded pieces of code.
Removed them and renamed various variables to better indicate their use
within the current impl.
Added a wrapper around the callafter to ensure it is not called if the
supplied file is intended to be edited despite not existing.
This is a minor issue but does cause an error to be thrown so we check
for it here and just dont run the callafter function if there is an
issue.
Removed the autocmd which was previously used to update the autocomplete
as this has been outsourced to telescope.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>