Fixed get_dirs to include subdirs
Previous impl would fail to recurse to subdirs below 1 level of depth.
Fixed impl to account for this and recurse down to ensure subdirs are
also added to the list of valid dirs.
This was not previously an issue as the pre-telescope impl would
continuously update the list of dirs with several calls to get_dirs to
refine the list as the user added additional dirs.
Since telescope gets the users response in one shot and does not update
the list it shown by default the new impl must be used.

Signed-off-by: Sean Allesina-McGrory <allesis@wwu.edu>